[operator] Improved Retry Queue logic in InformerController (#41)
Currently, in `operator.InformerController`, a user can set a
`RetryPolicy` func that will dictate when a failed watcher event should
be retried. However, any subsequent actions taken on that object will
dequeue any outstanding retries. This is ideal in the default case,
where a new update supersedes the old one, but has conflicts with the
`OpinionatedWatcher`, as updates which bump the ResourceVersion, but not
the Generation, are never received by the underlying update code
(`OpinionatedWatcher` drops them and reports back a success). That means
if an update fails and enters the retry queue, but a subsequent update
which changes non-spec content is received, the outstanding retry for
the original update, _which still needs to be performed_, is dropped,
and the update code behind the `OpinionatedWatcher` is never actually
called, making reconciliation impossible.

There are additional edge cases where a user may not wish to drop
outstanding retries, such as when an add is being retried and an update
comes in, but update and add do distinct enough things that both
handlers should still be called.

This PR addresses this issue by adding an optional exported field in the
`InformerController`: `RetryDequeuePolicy`. When nil, all outstanding
retries are dequeued the same as normal, but when present, a user can
dictate if a retry should be dequeued or not. The function is passed
information about the current action and the retry action to possibly
dequeue.

An "opinionated" `RetryDequeuePolicy` is also made available for a user,
which follows logic that gels with the `OpinionatedWatcher`, but also
allows for concurrent retries which are for different actions (i.e. an
add and an update).

Some changes were made to the retry logic to handle multiple retries for
the same object (namely moving from a `xsync.MapOf` to `ListMap`), but
the core logic of how retries works remains the same.