Admission Control (#29)
This PR introduces methods of doing validating and mutating admission
control via kubernetes webhooks, expressed in a generic manner. It
focuses on four parts:
1. A generic set of interfaces and structs for admission control (in the
`resource` package)
2. A kubernetes webhook server which can run as a controller, and
consumes handlers for mutating/validating kinds
3. Opinionated mutating/validating controllers which handle essential
grafana kind metadata for you
4. `project local generate` updates to allow for the generated manifests
to include webhooks. This includes some changes to the Tiltfile to make
sure we don't run into weird blocking scenarios for startup.

(2) is the largest component of this, and most of the complexity is
around (a) translating kubernetes, and (b) associating specific
mutating/validating controllers with kinds. Translating kubernetes is of
particular complexity, because we need to _also_ handle limitations in
JSONPatch w.r.t. kubernetes when it comes to annotations--in essence, we
need to modify the patch operations returned to us by the controllers
based on whether or not it has non-kubernetes metadata which must be
encoded in annotations, and if annotations is currently absent/an empty
map. This is a large PR, though more than half of the lines of code
added should be from tests.

I'm still not 100% happy with the implementation in the `k8s` package,
but I think the exposed interfaces/types are fine going forward, and we
may just need to tinker with the logic in the functions themselves, so
it would be transparent to users.

---------

Co-authored-by: Igor Suleymanov <radiohead@users.noreply.github.com>