[bugfix] Translate RawToObject fix for newer Thema versions (#67)
Thema in generated code was unhappy with the unmarshal call missing a
few fields in the metadata. Made sure those fields are present. This is
a short-term fix, with a better, full fix that should be done to make
sure the complete metadata object is passed to the Unmarshal call,
rather than being fragmented between Unmarshal and SetCommonMetadata.