Logging (#89)
# What this PR does
This PR adds the `logging` package to the repo, which contains a
`Logger` interface definition, an `slog` implementation of that
interface, methods to put a Logger into the context/retrieve it from the
context, and a DefaultLogger which is used when either references
directly, or no Logger is in the interface (defaults to a No-op logger).

It also introduces using logging in a few places, and a
`router.NewLoggingMiddleware` which logs each request, and
`plugin.PluginLogger`, which wraps a plugin-sdk-go `log.Logger` to be
compliant with `logging.Logger`, and add `traceID` to each message
logged with a context.

Generated project code now initializes the DefaultLogger in the plugin
to a `plugin.PluginLogger` wrapping plugin-sdk-go's `log.DefaultLogger`
(with a `pluginID` field attached), and in the operator to the slog
implementation.

Local env setup now allows for specifying a `loki` data source, which
gets set-up to monitor pod logs in the local kubernetes cluster.

## Breaking Changes
* `operator.InformerController.ErrorHandler` function now takes both a
`context.Context` and an `error`, as opposed to previously just taking
an `error`. The default handler will now log the error.
* `router.NewJSONRouter` and `router.NewJSONRouterWithErrorHandler` no
longer take a `log.Logger` parameter, and will use the `logging.Logger`
from the context (or `logging.DefaultLogger`) to log errors.

## Other Updates
The golang version in `go.mod` has been updated to 1.21, as that
introduced `log/slog` to the standard library.

The golangci linter needed to be updated to handle go 1.21. Once
updated, a few minor updates had to be made to unrelated files based on
new enforcement (namely `interface{}` -> `any`). The GH action linter
has also been updated, but I'm not sure if it will run the new one on
the PR branch.


Part of https://github.com/grafana/grafana-app-sdk/issues/24

---------

Co-authored-by: Igor Suleymanov <radiohead@users.noreply.github.com>