Metrics Instrumentation (#103)
# What This PR Does

Instruments the app-sdk for prometheus metrics, and adds local setup for
scraping said metrics.

## Metrics Instrumentation

The `k8s` and `operator` packages are instrumented with prometheus
collectors which can use used to emit metrics. Additionally, a metrics
middleware was added to the `router` package for metrics instrumentation
in routing.

Additionally, a `metrics` package has been added which includes an
`Exporter`, which can be used to register the collectors returned by the
`PrometheusCollectors()` methods on the instrumented types in the sdk,
and can also be run as a controller in the operator to expose the
`/metrics` endpoint for scraping.

### k8s

`ClientGenerator` and `SchemalessClient` use the new `MetricsConfig`
part of the `ClientConfig` to create two collectors that are used for
all kubernetes client operations:
* `kubernetes_client_request_duration_seconds` (Histogram) records the duration
of each kubernetes request, with labels for `status_code`, `method`, and
`path`
* `kubernetes_client_requests_total` (Counter) increments on every new
kubernetes request made by a client, with labels for `status_code`,
`method`, and `path`

### operator

The only object in the `operator` package instrumented for metrics is
the `InformerController`. The introduction of metrics also introduces a
minor **breaking change**, in that the `NewInformerController` now
requires an argument of an `InformerControllerConfig` (which contains
metrics configuration). The collectors created and used by the
`InformerController` all have labels for `event_type` (using the
`ReconcileAction` type constants) and `kind` (the kubernetes kind
string):
* `reconcile_duration_seconds` (Histogram) records the total duration of
each event's reconcile process, that is, the time from when the event is
received by the controller to when the controller is finished processing
all associated watchers and reconcilers for it
* `informer_events_total` (Counter) increments on every event received
by the informer controller
* `ongoing_reconcile_events` (Gauge) increments when an event is
received by the controller and decrements when it has completed
processing
* `reconciler_process_duration_seconds` (Histogram) records the duration
of each reconciler's `Reconcile()` call. This is distinct from
`reconcile_duration_seconds` as it records for _each_ reconciler that is
in the controller. Even with oly one reconciler in the controller, it
cuts out the overhead from "event received" to "event finished
processing," only looking at the duration of the `Reconcile()` call
itself.
* `watcher_process_duration_seconds` (Histogram) records the duration of
each `Add`/`Update`/`Delete` Watcher call. This is semantically similar
to `reconciler_process_duration_seconds`, but for watchers.
* `ongoing_reconcile_processes` (Gauge) increments for each reconciler's
`Reconcile()` call, or each Watcher's `Add`/`Update`/`Delete` call, and
decrements when they finish. This is the watcher/reconciler-specific
variant of `ongoing_reconcile_events`, much like
`reconciler_process_duration_seconds` and
`watcher_process_duration_seconds` are the reconciler/watcher-specific
variants of `reconcile_duration_seconds`

### router

`NewMetricsMiddleware` creates a new middleware which records metrics
about requests. These metrics are the same as used in the
WeaveWorks/common server for consistency, have the labels `status_code`,
`method`, and `route` (which is the matched route name in the router,
not the path, just like what is used in the otel trace), and consist of:
* `request_duration_seconds` (Histogram) records the request duration,
from the second the middleware is hit to the time the response is
returned to the middleware.
* `requests_total` (Counter) increments on every new request
* `request_message_bytes` (Histogram) records the size of the request
body
* `response_message_bytes` (Histogram) records the size of the response
body
* `inflight_requests` (Gauge) increments when a request hits the
middleware, and decrements when the middleware finishes

## Codegen

Codegen has been updated for both plugin and operator to include metrics
setup. For the plugin, this just entails registering the
`k8s.ClientRegistry` collectors with the default prometheus Registerer.
For the operator, it includes setting up a `metrics.Exporter` and
registering metrics with it, then running it as part of the operator.

## Local Environment

Local environment setup has been updated to include agent scraping of
the grafana deployment (and the plugin's metrics as part of that
deployment), and the operator's `/metrics` endpoint. This is currently
not very configurable, and will likely be made more configurable in a
future PR.