Separate into modules for plugin and CLI to avoid versioning issues (#190)
This PR introduces extra go submodules for the `cmd/grafana-app-sdk` and
`plugin` directories, and a `go.work` to allow them to incorporate
properly. It also moves the `codegen`, `kindsys`, and `internal`
packages into the `cmd/grafana-app-sdk` submodule. This accomplishes two
major things:
1. The CLI and library dependencies are no longer "mixed." The CLI still
have some upward dependencies on the`resource` and `k8s` packages, but
now using any of the libraries (`resource`, `k8s`, `operator`, etc.)
does not require the dependencies of the `codegen` package or CLI. The
`plugin` package has been similarly isolated because dependency versions
in the root module will overwrite the ones in the submodule, and the
`codegen` package requires an older version of `getkin/kin-openapi` than
`plugin` requires.
2. Removes the need for the `replace` directive version pinning, which
was preventing `go install` from working correctly, and _also_ required
that a user of the app-sdk as a library also provided that `replace`
directive in their `go.mod`.

Resolves https://github.com/grafana/grafana-app-sdk/issues/189