Refactor of the Object interface to be a superset of the kubernetes interfaces, introduce Kinds (#200)
This PR supersedes https://github.com/grafana/grafana-app-sdk/pull/182
and is the first move toward the actual implementation of the ideas
present in that PR. Reviewers of
https://github.com/grafana/grafana-app-sdk/pull/182 have also been added
to this PR. *EDIT: I am unable to add some reviewers to this PR for
unexplained GitHub reasons*

This PR is a pull into a branch which is not main, as there will be
several PRs to get this into a working (compiling) shape, then that
branch will be pulled into main. All the work has been done to get this
into a working shape already, but the PR for that would be far too
large, so it is being split into these component PRs, to allow for
reviews and comments of each part of the design/implementation process.
This PR is _just_ the `resource` package, which introduces:
* Changes to the `Object` interface to make it a superset of the
kubernetes `runtime.Object`, `metav1.Object`, and `schema.ObjectKind`,
along with additional methods.
* Changes to the `ListObject` interface to also make it a superset of
`runtime.Object`, `schema.ObjectKind`, and `metav1.ListInterface`
* Removal of the Simple and Store objects in favor of a few basic
`Object` implementations: `UntypedObject`, which contains unstructured
spec and subresource data; `TypedSpecObject`, which contains typed spec
data, `TypedSpecStatusObject`, which contains typed spec and status
data; and `TypedObject`, which contains typed spec data and a typed
catalog of subresources.
* Since the `Object` interface no longer has the `Unmarshal` method,
introduce a `Codec` interface for reading/writing `Object` interfaces,
and a `Kind` struct type which unifies a `Schema` and one or more
`Codec`s based on the kind encoding. Anything which needs to read/write
bytes (`ClientRegistry`, and things which use it) now accept the `Kind`
as a parameter instead of `Schema`.
* ~Deprecation of `Schema` and its associated implementations and groups
in favor of `Kind`, which is `Schema` + an encoder and decoder for the
object.~
* ~Simple implementations of `Kind`: `TypedKind` and `UntypedKind`~
* Updates to the various stores to deal with these changes

There are a few open questions for reviewers noted in `TODO` comments,
namely around whether we need certain methods, or if certain method
signatures should be expanded upon. Other feedback is more than welcome.

---------

Co-authored-by: Igor Suleymanov <radiohead@users.noreply.github.com>