Fix generated files' permissions
When using `grafana-app-sdk project init` files are being created with
the executable bit set in cases where it's not needed (e.g. some code,
makefiles, etc). Given how there are two different functions,
`writeFile` and `writeExecutableFile`, it's clear that this was not the
intention.

Modify `writeFile` so that it creates files with the read/write bits
set, but not the executuable one.

Modify both `writeFile` and `writeExecutableFile` so that the process'
umask is respected (i.e. pass 0666 instead of 0644).

Signed-off-by: Marcelo E. Magallon <marcelo.magallon@grafana.com>