Update application design documentation
WHAT

This commit updates the documentation related to the application design.
The new version is more up-to-date, with explanations of the underlying
concepts and updated application types.

WHY

To make sure that our documentation is in line with the current vision
and development direction of the app platform and the app SDK.

Signed-off-by: Igor Suleymanov <igor.suleymanov@grafana.com>