[docs] Update Operator Docs (#269)
Rather large docs enhancement, update the operator doc and add a new doc
on Writing an Operator, which breaks down the `simple.Operator` and how
to use more complex behavior with `operator.Operator`. As part of this
enhancement, additions were made to the tutorial doc for adding
functionality to the watcher and adding admission control to the
operator. Some small fixes to local project setup and codegen were made
along the way as a product of testing the new tutorial changes.

Resolves https://github.com/grafana/grafana-app-sdk/issues/255

A future PR should be made to introduce multiple versions to the
tutorial to go alongside
https://github.com/grafana/grafana-app-sdk/blob/main/docs/custom-kinds/managing-multiple-versions.md,
but that is out-of-scope for this PR.