Pagination in Store and TypedStore (#290)
Auto-pagination in Store and TypedStore, with additional methods for
fetching a specific page without using auto-pagination.

To allow for customizing the perPage limit, an additional parameter was
added to the `List` method. It will automatically fetch each page and
concatenate into a single response. To fetch a single page at a time, a
new method called `ListPage` has been introduced.

---------

Co-authored-by: Igor Suleymanov <radiohead@users.noreply.github.com>