[codegen] Test Improvements (#300)
Improve testing for codegen by testing all go and TypeScript codegen
paths. Includes many more test files, and a few drive-by fixes
discovered while improving the tests. The largest drive-by fix is for
multiple versions of a CUE kind, where the versions are in a map and
added to a slice, so the order in the slice is random (go enforces
random ordering of keys when iterating through a map). Slice elements
are now sorted by attempting to use either kubernetes or thema
versioning, falling back to string compare if neither format matches.

Nearly all the changes here are adding new generated files to
`golden_generated`.