Update module and workspace dependency files
WHAT

This commit updates the *.mod and *.sum files to make sure that they are
up-to-date with the status of the dependencies.

The changes have been generated by running `make`.

WHY

It looks like https://github.com/grafana/grafana-app-sdk/pull/343 has
somehow left the dependencies in an incorrect state and now all PRs are
failing due to inconsistent state.

Signed-off-by: Igor Suleymanov <igor.suleymanov@grafana.com>