[operator] Custom Cache Informer (#336)
## What this PR does

This PR introduces a new `operator.Informer`-implementing type called
`operator.CustomCacheInformer`. This new informer type functions
similarly to `operator.KubernetesBasedInformer`, but allows a user to
specify a custom `cache.Store` implementation to use for storing
resources retrieves from the `ListerWatcher`, rather than only using an
in-memory map as `operator.KubernetesBasedInformer` does.

This had to be implemented as a separate Informer as the kubernetes
informer(s) do not let you specify the `cache.Store` to use as the
client state value, as it always initialized a new `cache.DeltaFIFO` on
`Run()`
([source](https://github.com/kubernetes/client-go/blob/master/tools/cache/shared_informer.go#L471-L475)),
and
[NewSharedIndexInformerWithOptions](https://github.com/kubernetes/client-go/blob/master/tools/cache/shared_informer.go#L268)
only allows the caller to provide indexers that are used to create the
`cache.Indexer` used by the DeltaFIFO.

The new informer in this PR replicates (with modifications where
appropriate) the unexported functionality in `cache.sharedIndexInformer`
needed for the informer to run, using exported kubernetes code where
available.

### Memcached `cache.Store`

This PR also contains an implementation of `cache.Store` using
memcached. Several `cache.Store` methods cannot be done via memcached,
but the only one that is used by the informer is `ListKeys()`. This is
optionally implemented in the memcached implementation of `cache.Store`
by keeping an in-memory map of keys, updated whenever a key is read or
written.

The memcached store always internally prefixes keys with the kind, in
order to avoid conflicts when using the same memcached instance(s) for
multiple kinds for the same (or multiple) operators.

A side-effect of the externalization of the cache is that the initial
list no longer triggers an `Add` event on existing resources (as they
are already in the cache which now persists across restarts). This means
that users of the `OpinionatedWatcher` and `OpinionatedReconciler` will
not receive any events at startup unless they are a resource which does
not exist in the cache, or the cache state for the resource is different
from the API state (if they are the same, an `Update` event is still
issued, but the Opinionated variants drop it as the RV has not changed).
If a user does not use the opinionated variants, they will see Update
events with no change in RV for resources that already exist in the
cache and have not been changed in the API server, rather than the Add
events they previously received on startup when the cache was always
empty by design.

## Testing

An application written with a CustomCacheInformer and MemcachedStore
were tested in a k3d cluster using the docker memcached image as the
memcached store. There has not been significant load testing done in the
new informer nor the MemcachedStore. New unit tests were introduced for
the CustomCacheInformer which test minor loads, but are not real-world
scenarios, and do not test end-to-end, only from event ingestion via the
ListerWatcher.

Resolves https://github.com/grafana/grafana-app-sdk/issues/263

---------

Co-authored-by: Igor Suleymanov <radiohead@users.noreply.github.com>