Fix Integration Tests Workflow and Retract Bad Tags (#386)
Fix the integration tests step in the release workflow by giving the
adjusted path for the `go.mod` (tested [in this
fork](https://github.com/IfSentient/grafana-app-sdk/actions/runs/10692236317)).

Also retract the release that tried to build with this error (no release
binaries), and the previous release (v0.18.3) as that tag is mismatched
between GOPROXY and here. [Retraction is the correct way to remove a tag
per go](https://go.dev/ref/mod#go-mod-file-retract), so the v0.18.3 tag
being deleted and replaced was incorrect behavior and should be listed
in the retracted list in `go.mod` in both main and `plugin` modules.