[AppManifest] Add Schema Information, CUE Definition (#426)
This PR adds kind version schemas to `app.ManifestData`, replacing the
`any` type with a `VersionSchema` which marshals/unmarshals as CRD-like
OpenAPI, but allows for extracing schema information as
getkin-kin-openapi3, or (in the future) kube-openapi.

This PR also adds a CUE definition for the AppManifest to match how the
CRD will be presented in an app platform API server. The definition is
in the format of a resource defined for an app platform application.