[codegen][docs] Update New Project Codegen and Tutorial Doc (#423)
This PR updates the new project codegen from `grafana-app-sdk project
component add operator` to use `simple.App` and `operator.Runner`, as
opposed to `simple.Operator`. It now splits the app out into two files:
`/cmd/operator/main.go`, which contains the loading of config, and
running the operator runner, and `/pkg/app/app.go`, which contains the
app logic, including the Provider and NewApp functionality. This allows
the app to be re-used for other runners if desired, though other
`/cmd/*` packages.

This PR also updates the tutorial based on these changes, so that it
follows along with the new way of writing apps and reflects the codegen
as of this PR.

Finally, there is also an update to the project local generation, used
with `grafana-app-sdk project local generate`, which now no longer uses
values in the config for admission/conversion capabilities and instead
uses the capabilities declared in the app manifest.