Allow project local generate to work with no kinds defined (#461)
Motivation: I wanted to adopt the grafana-app-sdk tooling and structure
for an existing plugin because we have plans to introduce some custom
Kinds there. However, in the spirit of smaller changes, I wanted to
restructure the project without adding custom Kinds. [You can see my
example here](https://github.com/grafana/grafana-pdc-app/pull/541).

This PR shows what I needed to be able to make use of `grafana-app-sdk
project local generate` when I didn't have any custom kinds defined.

I couldn't see a nice way to refactor the cue generator to gracefully
handle the lack of a cue module (and it didn't feel right), so I added
this `none` format which indicates that we don't have any kinds defined.

Happy to be guided on this, I mainly made this PR as a discussion point.
But I'd love to be able to have this functionality.