[CLI][codegen] Use `yarn create @grafana/plugin` to Generate Front-End Code (#491)
Instead of using a snapshot of the output of `yarn create
@grafana/plugin` with some small tweaks when adding the `frontend`
component with `grafana-app-sdk project component add`, invoke the
`yarn` command directly, then tweak the output as needed. This keeps us
from falling behind the current plugin bootstrapping. As yarn is
required to build the frontend already, we have the component add check
if yarn is installed, and error if not, requesting that the user must
have yarn installed to add the frontend component.

Resolves https://github.com/grafana/grafana-app-sdk/issues/473

---------

Co-authored-by: Igor Suleymanov <radiohead@users.noreply.github.com>