[LTS v0.24][OpinionatedWatcher/Reconciler] Patch Against the Preferred Version (#520)
# What this PR Does

This is a duplicate of
https://github.com/grafana/grafana-app-sdk/pull/511, applied to the
`lts/v0.24` branch instead of `main`, to be a part of the next LTS
release.

### Original PR Text

To allow the OpinionatedWatcher and OpinionatedReconciler to issue patch
requests to the preferred API version, instead of the version being
watched (as conversion for the patching can lead to edge-case conversion
problems), introduce `k8s.DynamicPatcher`, which issues patch requests
to the preferred API version of a GroupKind, using the discovery client
to get (and cache) preferred version(s).

This change also introduces a wrapper for this client to avoid changing
the `operator.PatchClient` used for OpinionatedWatcher and
OpinionatedReconciler--this wrapper is unexported and used in
`simple.App` and `simple.Operator` so that apps built that way will work
properly. This is so the same change can also be made to the `lts/v0.24`
branch. A later state of this will likely move the DynamicPatcher to
only allow for patching of metadata, or at least having the opinionated
watcher and reconciler use an interface that only patches metadata,
rather than the whole object, which the DynamicPatcher can satisfy.

Relates to https://github.com/grafana/grafana-app-sdk/issues/506